requirements.txt
openai==1.99.6
numpy==2.3.2
python-dotenv==1.1.1
streamlit==1.48.0
pydub==0.25.1
faiss-cpu==1.11.0